Remove base class inheritance from CommandlineOption test
